+++ frontend/src/components/ControlDots.tsx
