+++ frontend/src/pages/RollingpaperCreationPage/components/Step2.tsx
