+++ frontend/src/pages/RollingpaperCreationPage/index.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useRef, useState } from "react";
 import styled from "@emotion/styled";
 
+import ControlDots from "@/components/ControlDots";
+
 import useCreateMemberRollingpaper from "@/pages/RollingpaperCreationPage/hooks/useCreateMemberRolliingpaper";
 import useCreateTeamRollingpaper from "@/pages/RollingpaperCreationPage/hooks/useCreateTeamRollingpaper";
 
-import Step1 from "@/pages/RollingpaperCreationPage/steps/Step1";
-import Step2 from "@/pages/RollingpaperCreationPage/steps/Step2";
-import Step3 from "@/pages/RollingpaperCreationPage/steps/Step3";
-
-import ControlDots from "@/pages/RollingpaperCreationPage/components/ControlDots";
+import Step1 from "@/pages/RollingpaperCreationPage/components/Step1";
+import Step2 from "@/pages/RollingpaperCreationPage/components/Step2";
+import Step3 from "@/pages/RollingpaperCreationPage/components/Step3";
 
 import { Recipient, Rollingpaper, Team, TeamMember } from "@/types";
 import { RECIPIENT } from "@/constants";
